Add day 5
Add initial solution. This is a bit of brute force swapping bad pairs
until there is a valid solution. An alternative would be to do a breadth
first search, but it might cause issues if there are any diamonds in the
graph (and we just want to see if there are left to right matches).