Add day 9
This is pretty ugly. In short, I'm trying to do everything I can in both
parts to avoid having to store anything in memory. Instead, calculate
what ought to be printed at each location and update the list to keep
track of used values or blocks.

Most of the ugliness comes from having to track the same invariants
across each loop. A lot of that would be cleaner if the two main inner
loops (count forward/count backward) were pulled into separate
functions.