Add day 11
The insight here is that each stone is independent. Rather than
calculating the full array, just calculate the number of stones each
initial stone will contribute. Then memoize calls, since there should be
a lot of repetition.

I expected this to be a "big number" problem, so the time between parts
one and two was 16 seconds.