Add day 12
Part one was trivial. Part two was a little trickier. I suspect there is
an easier way to do this by counting turns directly, but it wasn't clear
how to generalize it to work with an arbitrarily ordered flood fill. The
idea here of finding a "representative" portion of the perimeter to
count as the side is an interesting technique.