Add day 14
This is tricky since the example doesn't really show the easter egg (as
far as I could tell). I initially found it by searching for a long line
of contiguous robots, but it was just luck. After, I tried checking if
all robots in unique locations also works, and that seems simpler.