Add day 16
Djikstra with a twist. For part two, the idea is to just pass along the
entire path with the search. This ends up being a little slow since
we're only excluding paths that are definitely sub-optimal, but still
have to let all paths finish and reprocess. An alternative to this would
be to keep a map from (row, col, dir) tuples to paths and keep updating
that set. That should be cleaner.