Add day 17
Part two was a little tricky. Initially, I was solving one value at a
time without searching the full range. The issue with this is that
the range might not end up being contiguous for a single output, so you
can get something that matches everything but the "first" value, and
then nothing in the range matches. To solve this, you have to look
through the full range.