Add day 20
This took some time because I over-complicated it. I assumed that a
cheat was only a single crossing of a barrier as opposed to just
disabling collision altogether (as it says in the problem). Without that
constraint, this problem is a lot simpler.

Some of the path length code could be simplified to remove the need to
track the max length, but it seems equally as complex, so this is fine
as is.