Add day 22
Surprisingly straightforward. The only clever parts are using bitwise
operations everywhere and finding a compact way to encode a sequence as
a single 20 bit value. I'm not sure this actually has a significant
effect since storing 4 integer values (or even 4 chars in a different
language) isn't significantly more. There are also only 19^4 sequences,
which is small, so this is a micro-optimization at best.