Add day 23
This is just finding a max-clique algorithm (Bron-Kerbosch) and
implementing it. The main constraints are that there are 520 vertices
and each node has 13 edges, so it's a little too big to just try all
combinations, but hard to tell if it was too big for this. It works
quickly enough though, so it seems fine. There's probably room to
improve this if we assume that the clique we're looking for has 13
nodes, but it's not obvious what that improvement would be.