Automate part 2
The overall process is to find each broken gate, find the name of the
carry output of the previous value, then find names of each output that
should be known. Specifically, it only identifies issues in gates where
an XOR gate is swapped (which happens to be the case for my input). In
theory, swaps without XOR gates should be impossible either because they
create a loop or because they would not be detectible.